reformatting the directory
