+++ source/FunctionReadingX/test/test_dict_function.py
@@ -1,8 +1,8 @@
-from source.FunctionReading.function_reading import AlgebraicFunctionReading as afr
+from source.FunctionReadingX.Function_Reading import AlgebraicFunctionReading as afr
 
 def test_error_symbol():
     try:
-        afr(list_function=[2,2.5,-2,'/'])
+        afr(list_function={'-':2})
         fail = False
     except:
         fail = True
@@ -10,7 +10,7 @@ def test_error_symbol():
     
 def test_error_str_number():
     try:
-        afr(list_function=['2.2','2'])
+        afr(list_function={'x':'2'})
         fail = False
     except:
         fail = True
@@ -18,7 +18,7 @@ def test_error_str_number():
     
 def test_error_list_max():
     try: # no pass
-        afr(list_function=[2,-1,2,4,5,0,10])
+        afr(list_function={'x6':3,'x4':4})
         fail = False
     except:
         fail = True
@@ -26,15 +26,23 @@ def test_error_list_max():
     
 def test_error_list_min():
     try: # no pass
-        afr(list_function=[])
+        afr(list_function={})
+        fail = False
+    except:
+        fail = True
+    assert fail == True
+    
+def test_error_values_empty():
+    try: # no pass
+        afr(list_function={'x2':' ','x':2})
         fail = False
     except:
         fail = True
     assert fail == True
     
 def test_pass_list_min():
-    try:
-        afr(list_function=[2])
+    try: # no pass
+        afr(list_function={'x':1})
         fail = False
     except:
         fail = True
@@ -42,7 +50,7 @@ def test_pass_list_min():
         
 def test_pass_list_max():
     try:
-        afr(list_function=[2,-1,2,4,5,0])
+        afr(list_function={'x5':1,'x4':1,'x3':4,'x2':-2,'x':1,'c':2})
         fail = False
     except:
         fail = True
