+++ source/FunctionReadingX/test/test_validation.py
@@ -1,5 +1,4 @@
-import pytest
-from source.FunctionReading.function_reading import AlgebraicFunctionReading as afr
+from source.FunctionReadingX.Function_Reading import AlgebraicFunctionReading as afr
 
 def test_validate_arg_error():
     try:
